Converted to Firebase as the backend which substantially simplifies the backend processing.  Now some tidying up to do with key values and data design of the timestamps
