--- CDCrew+CoreDataClass.swift
@@ -1,16 +1,0 @@
-//
-//  CDCrew+CoreDataClass.swift
-//  
-//
-//  Created by Paul Ventisei on 23/09/2018.
-//
-//
-
-import Foundation
-import CoreData
-
-
-public class CDCrew: NSManagedObject {
-    
-
-}
