--- CDEvent+CoreDataClass.swift
@@ -1,15 +1,0 @@
-//
-//  CDEvent+CoreDataClass.swift
-//  
-//
-//  Created by Paul Ventisei on 23/09/2018.
-//
-//
-
-import Foundation
-import CoreData
-
-
-public class CDEvent: NSManagedObject {
-
-}
