--- CDStage+CoreDataClass.swift
@@ -1,15 +1,0 @@
-//
-//  CDStage+CoreDataClass.swift
-//  
-//
-//  Created by Paul Ventisei on 23/09/2018.
-//
-//
-
-import Foundation
-import CoreData
-
-
-public class CDStage: NSManagedObject {
-
-}
