--- GTLRObservedtimes.h
@@ -1,14 +1,0 @@
-// NOTE: This file was generated by the ServiceGenerator.
-
-// ----------------------------------------------------------------------------
-// API:
-//   observedtimes/v1
-// Description:
-//   ObservedTimes API v1. This API allows the GET method to get a collection of
-//   observed times since the last time the user asked for a list and a POST
-//   method
-//   to record new times
-
-#import "GTLRObservedtimesObjects.h"
-#import "GTLRObservedtimesQuery.h"
-#import "GTLRObservedtimesService.h"
