--- GTLRObservedtimesObjects.m
@@ -1,200 +1,0 @@
-// NOTE: This file was generated by the ServiceGenerator.
-
-// ----------------------------------------------------------------------------
-// API:
-//   observedtimes/v1
-// Description:
-//   ObservedTimes API v1. This API allows the GET method to get a collection of
-//   observed times since the last time the user asked for a list and a POST
-//   method
-//   to record new times
-
-#import "GTLRObservedtimesObjects.h"
-
-// ----------------------------------------------------------------------------
-//
-//   GTLRObservedtimes_RowTimePackageClockSyncReply
-//
-
-@implementation GTLRObservedtimes_RowTimePackageClockSyncReply
-@dynamic clientTime, diffInMs, serverTime;
-
-+ (NSDictionary<NSString *, NSString *> *)propertyToJSONKeyMap {
-  NSDictionary<NSString *, NSString *> *map = @{
-    @"clientTime" : @"client_time",
-    @"diffInMs" : @"diff_in_ms",
-    @"serverTime" : @"server_time"
-  };
-  return map;
-}
-
-@end
-
-
-// ----------------------------------------------------------------------------
-//
-//   GTLRObservedtimes_RowTimePackageClockSyncRequest
-//
-
-@implementation GTLRObservedtimes_RowTimePackageClockSyncRequest
-@dynamic clientTime;
-
-+ (NSDictionary<NSString *, NSString *> *)propertyToJSONKeyMap {
-  return @{ @"clientTime" : @"client_time" };
-}
-
-@end
-
-
-// ----------------------------------------------------------------------------
-//
-//   GTLRObservedtimes_RowTimePackageCrew
-//
-
-@implementation GTLRObservedtimes_RowTimePackageCrew
-@dynamic category, cox, crewId, crewName, crewNumber, division, endTimeLocal,
-         picFile, rowerCount, stages, startTimeLocal;
-
-+ (NSDictionary<NSString *, NSString *> *)propertyToJSONKeyMap {
-  NSDictionary<NSString *, NSString *> *map = @{
-    @"crewId" : @"crew_id",
-    @"crewName" : @"crew_name",
-    @"crewNumber" : @"crew_number",
-    @"endTimeLocal" : @"end_time_local",
-    @"picFile" : @"pic_file",
-    @"rowerCount" : @"rower_count",
-    @"startTimeLocal" : @"start_time_local"
-  };
-  return map;
-}
-
-+ (NSDictionary<NSString *, Class> *)arrayPropertyToClassMap {
-  NSDictionary<NSString *, Class> *map = @{
-    @"stages" : [GTLRObservedtimes_RowTimePackageStage class]
-  };
-  return map;
-}
-
-@end
-
-
-// ----------------------------------------------------------------------------
-//
-//   GTLRObservedtimes_RowTimePackageCrewList
-//
-
-@implementation GTLRObservedtimes_RowTimePackageCrewList
-@dynamic crews, eventId;
-
-+ (NSDictionary<NSString *, NSString *> *)propertyToJSONKeyMap {
-  return @{ @"eventId" : @"event_id" };
-}
-
-+ (NSDictionary<NSString *, Class> *)arrayPropertyToClassMap {
-  NSDictionary<NSString *, Class> *map = @{
-    @"crews" : [GTLRObservedtimes_RowTimePackageCrew class]
-  };
-  return map;
-}
-
-@end
-
-
-// ----------------------------------------------------------------------------
-//
-//   GTLRObservedtimes_RowTimePackageEvent
-//
-
-@implementation GTLRObservedtimes_RowTimePackageEvent
-@dynamic eventDate, eventDesc, eventId, eventName;
-
-+ (NSDictionary<NSString *, NSString *> *)propertyToJSONKeyMap {
-  NSDictionary<NSString *, NSString *> *map = @{
-    @"eventDate" : @"event_date",
-    @"eventDesc" : @"event_desc",
-    @"eventId" : @"event_id",
-    @"eventName" : @"event_name"
-  };
-  return map;
-}
-
-@end
-
-
-// ----------------------------------------------------------------------------
-//
-//   GTLRObservedtimes_RowTimePackageEventList
-//
-
-@implementation GTLRObservedtimes_RowTimePackageEventList
-@dynamic events;
-
-+ (NSDictionary<NSString *, Class> *)arrayPropertyToClassMap {
-  NSDictionary<NSString *, Class> *map = @{
-    @"events" : [GTLRObservedtimes_RowTimePackageEvent class]
-  };
-  return map;
-}
-
-@end
-
-
-// ----------------------------------------------------------------------------
-//
-//   GTLRObservedtimes_RowTimePackageObservedTime
-//
-
-@implementation GTLRObservedtimes_RowTimePackageObservedTime
-@dynamic crew, eventId, obsType, stage, time, timeId, timestamp;
-
-+ (NSDictionary<NSString *, NSString *> *)propertyToJSONKeyMap {
-  NSDictionary<NSString *, NSString *> *map = @{
-    @"eventId" : @"event_id",
-    @"obsType" : @"obs_type",
-    @"timeId" : @"time_id"
-  };
-  return map;
-}
-
-@end
-
-
-// ----------------------------------------------------------------------------
-//
-//   GTLRObservedtimes_RowTimePackageObservedTimeList
-//
-
-@implementation GTLRObservedtimes_RowTimePackageObservedTimeList
-@dynamic eventId, lastTimestamp, times;
-
-+ (NSDictionary<NSString *, NSString *> *)propertyToJSONKeyMap {
-  NSDictionary<NSString *, NSString *> *map = @{
-    @"eventId" : @"event_id",
-    @"lastTimestamp" : @"last_timestamp"
-  };
-  return map;
-}
-
-+ (NSDictionary<NSString *, Class> *)arrayPropertyToClassMap {
-  NSDictionary<NSString *, Class> *map = @{
-    @"times" : [GTLRObservedtimes_RowTimePackageObservedTime class]
-  };
-  return map;
-}
-
-@end
-
-
-// ----------------------------------------------------------------------------
-//
-//   GTLRObservedtimes_RowTimePackageStage
-//
-
-@implementation GTLRObservedtimes_RowTimePackageStage
-@dynamic label, stageIndex;
-
-+ (NSDictionary<NSString *, NSString *> *)propertyToJSONKeyMap {
-  return @{ @"stageIndex" : @"stage_index" };
-}
-
-@end
