--- GTLRObservedtimesQuery.h
@@ -1,163 +1,0 @@
-// NOTE: This file was generated by the ServiceGenerator.
-
-// ----------------------------------------------------------------------------
-// API:
-//   observedtimes/v1
-// Description:
-//   ObservedTimes API v1. This API allows the GET method to get a collection of
-//   observed times since the last time the user asked for a list and a POST
-//   method
-//   to record new times
-
-#if GTLR_BUILT_AS_FRAMEWORK
-  #import "GTLR/GTLRQuery.h"
-#else
-  #import "GTLRQuery.h"
-#endif
-
-#if GTLR_RUNTIME_VERSION != 3000
-#error This file was generated by a different version of ServiceGenerator which is incompatible with this GTLR library source.
-#endif
-
-@class GTLRObservedtimes_RowTimePackageClockSyncRequest;
-@class GTLRObservedtimes_RowTimePackageObservedTime;
-
-// Generated comments include content from the discovery document; avoid them
-// causing warnings since clang's checks are some what arbitrary.
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wdocumentation"
-
-NS_ASSUME_NONNULL_BEGIN
-
-/**
- *  Parent class for other Observedtimes query classes.
- */
-@interface GTLRObservedtimesQuery : GTLRQuery
-
-/** Selector specifying which fields to include in a partial response. */
-@property(nonatomic, copy, nullable) NSString *fields;
-
-@end
-
-/**
- *  GTLRObservedtimesQuery_ClockClockcheck
- *
- *  Method: observedtimes.clock.clockcheck
- *
- *  Authorization scope(s):
- *    @c kGTLRAuthScopeObservedtimesUserinfoEmail
- */
-@interface GTLRObservedtimesQuery_ClockClockcheck : GTLRObservedtimesQuery
-// Previous library name was
-//   +[GTLQueryObservedtimes queryForClockClockcheckWithObject:]
-
-/**
- *  Fetches a @c GTLRObservedtimes_RowTimePackageClockSyncReply.
- *
- *  @param object The @c GTLRObservedtimes_RowTimePackageClockSyncRequest to
- *    include in the query.
- *
- *  @return GTLRObservedtimesQuery_ClockClockcheck
- */
-+ (instancetype)queryWithObject:(GTLRObservedtimes_RowTimePackageClockSyncRequest *)object;
-
-@end
-
-/**
- *  GTLRObservedtimesQuery_CrewList
- *
- *  Method: observedtimes.crew.list
- *
- *  Authorization scope(s):
- *    @c kGTLRAuthScopeObservedtimesUserinfoEmail
- */
-@interface GTLRObservedtimesQuery_CrewList : GTLRObservedtimesQuery
-// Previous library name was
-//   +[GTLQueryObservedtimes queryForCrewList]
-
-@property(nonatomic, copy, nullable) NSString *eventId;
-
-/**
- *  Fetches a @c GTLRObservedtimes_RowTimePackageCrewList.
- *
- *  @return GTLRObservedtimesQuery_CrewList
- */
-+ (instancetype)query;
-
-@end
-
-/**
- *  GTLRObservedtimesQuery_EventList
- *
- *  Method: observedtimes.event.list
- *
- *  Authorization scope(s):
- *    @c kGTLRAuthScopeObservedtimesUserinfoEmail
- */
-@interface GTLRObservedtimesQuery_EventList : GTLRObservedtimesQuery
-// Previous library name was
-//   +[GTLQueryObservedtimes queryForEventList]
-
-@property(nonatomic, copy, nullable) NSString *searchString;
-
-/**
- *  Fetches a @c GTLRObservedtimes_RowTimePackageEventList.
- *
- *  @return GTLRObservedtimesQuery_EventList
- */
-+ (instancetype)query;
-
-@end
-
-/**
- *  GTLRObservedtimesQuery_TimesListtimes
- *
- *  Method: observedtimes.times.listtimes
- *
- *  Authorization scope(s):
- *    @c kGTLRAuthScopeObservedtimesUserinfoEmail
- */
-@interface GTLRObservedtimesQuery_TimesListtimes : GTLRObservedtimesQuery
-// Previous library name was
-//   +[GTLQueryObservedtimes queryForTimesListtimes]
-
-@property(nonatomic, copy, nullable) NSString *eventId;
-
-@property(nonatomic, copy, nullable) NSString *lastTimestamp;
-
-/**
- *  Fetches a @c GTLRObservedtimes_RowTimePackageObservedTimeList.
- *
- *  @return GTLRObservedtimesQuery_TimesListtimes
- */
-+ (instancetype)query;
-
-@end
-
-/**
- *  GTLRObservedtimesQuery_TimesTimecreate
- *
- *  Method: observedtimes.times.timecreate
- *
- *  Authorization scope(s):
- *    @c kGTLRAuthScopeObservedtimesUserinfoEmail
- */
-@interface GTLRObservedtimesQuery_TimesTimecreate : GTLRObservedtimesQuery
-// Previous library name was
-//   +[GTLQueryObservedtimes queryForTimesTimecreateWithObject:]
-
-/**
- *  Fetches a @c GTLRObservedtimes_RowTimePackageObservedTime.
- *
- *  @param object The @c GTLRObservedtimes_RowTimePackageObservedTime to include
- *    in the query.
- *
- *  @return GTLRObservedtimesQuery_TimesTimecreate
- */
-+ (instancetype)queryWithObject:(GTLRObservedtimes_RowTimePackageObservedTime *)object;
-
-@end
-
-NS_ASSUME_NONNULL_END
-
-#pragma clang diagnostic pop
