--- GTLRObservedtimesService.h
@@ -1,80 +1,0 @@
-// NOTE: This file was generated by the ServiceGenerator.
-
-// ----------------------------------------------------------------------------
-// API:
-//   observedtimes/v1
-// Description:
-//   ObservedTimes API v1. This API allows the GET method to get a collection of
-//   observed times since the last time the user asked for a list and a POST
-//   method
-//   to record new times
-
-#if GTLR_BUILT_AS_FRAMEWORK
-  #import "GTLR/GTLRService.h"
-#else
-  #import "GTLRService.h"
-#endif
-
-#if GTLR_RUNTIME_VERSION != 3000
-#error This file was generated by a different version of ServiceGenerator which is incompatible with this GTLR library source.
-#endif
-
-// Generated comments include content from the discovery document; avoid them
-// causing warnings since clang's checks are some what arbitrary.
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wdocumentation"
-
-NS_ASSUME_NONNULL_BEGIN
-
-// ----------------------------------------------------------------------------
-// Authorization scope
-
-/**
- *  Authorization scope: View your email address
- *
- *  Value "https://www.googleapis.com/auth/userinfo.email"
- */
-GTLR_EXTERN NSString * const kGTLRAuthScopeObservedtimesUserinfoEmail;
-
-// ----------------------------------------------------------------------------
-//   GTLRObservedtimesService
-//
-
-/**
- *  Service for executing (null) queries.
- *
- *  ObservedTimes API v1. This API allows the GET method to get a collection of
- *  observed times since the last time the user asked for a list and a POST
- *  method
- *  to record new times
- */
-@interface GTLRObservedtimesService : GTLRService
-
-// No new methods
-
-// Clients should create a standard query with any of the class methods in
-// GTLRObservedtimesQuery.h. The query can the be sent with GTLRService's
-// execute methods,
-//
-//   - (GTLRServiceTicket *)executeQuery:(GTLRQuery *)query
-//                     completionHandler:(void (^)(GTLRServiceTicket *ticket,
-//                                                 id object, NSError *error))handler;
-// or
-//   - (GTLRServiceTicket *)executeQuery:(GTLRQuery *)query
-//                              delegate:(id)delegate
-//                     didFinishSelector:(SEL)finishedSelector;
-//
-// where finishedSelector has a signature of:
-//
-//   - (void)serviceTicket:(GTLRServiceTicket *)ticket
-//      finishedWithObject:(id)object
-//                   error:(NSError *)error;
-//
-// The object passed to the completion handler or delegate method
-// is a subclass of GTLRObject, determined by the query method executed.
-
-@end
-
-NS_ASSUME_NONNULL_END
-
-#pragma clang diagnostic pop
