+++ GoogleService-Info.plist
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>CLIENT_ID</key>
+	<string>664783221367-t5a6lqqdflsain89lp50osbth2lif7h3.apps.googleusercontent.com</string>
+	<key>REVERSED_CLIENT_ID</key>
+	<string>com.googleusercontent.apps.664783221367-t5a6lqqdflsain89lp50osbth2lif7h3</string>
+	<key>API_KEY</key>
+	<string>AIzaSyB0sx2TvVgOVsByaKsnl3HNB_nCgv3Jrpw</string>
+	<key>GCM_SENDER_ID</key>
+	<string>664783221367</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>com.rowtime.RowTime</string>
+	<key>PROJECT_ID</key>
+	<string>rowtime-firestore</string>
+	<key>STORAGE_BUCKET</key>
+	<string>rowtime-firestore.appspot.com</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:664783221367:ios:a38a905df0c481b5</string>
+	<key>DATABASE_URL</key>
+	<string>https://rowtime-firestore.firebaseio.com</string>
+</dict>
+</plist>
