--- RowTime-Bridging-Header copy.h
@@ -1,14 +1,0 @@
-//
-//  Use this file to import your target's public headers that you would like to expose to Swift.
-//
-
-#import <GTMSessionFetcher/GTMSessionFetcher.h>
-#import <GTMSessionFetcher/GTMSessionFetcherService.h>
-#import "GTLRObservedtimesQuery.h"
-#import "GTLRObservedtimesObjects.h"
-#import "GTLRObservedtimes.h"
-#import "GTLRObservedtimesService.h"
-
-
-
-
