+++ RowTime.xcodeproj/xcuserdata/paulventisei.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -7,7 +7,7 @@
 		<key>RowTime.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>0</integer>
+			<integer>19</integer>
 		</dict>
 	</dict>
 </dict>
