+++ RowTime.xcworkspace/contents.xcworkspacedata
@@ -2,9 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:RowTime-Bridging-Header.h">
-   </FileRef>
-   <FileRef
       location = "group:RowTime.xcodeproj">
    </FileRef>
    <FileRef
